Rainfall Prediction Mehsana District
Rainfall Prediction is the application of science and
technology to foretell the state of the atmosphere. It is
important to exactly determine the rainfall for effective use of
water resources, crop productivity and pre planning of water
structures
Rainfall is one of the most important hydrological parameter
on which the economy of India depends.
Monsoon prediction is clearly of great importance for India.